Conflicts:
	app/code/local/Tim/Recommendation/Block/Adminhtml/NoteReport/Grid.php
	app/code/local/Tim/Recommendation/Block/Adminhtml/Render/Average.php
	app/code/local/Tim/Recommendation/Block/Adminhtml/Render/DisplayNote.php
	app/code/local/Tim/Recommendation/Block/Adminhtml/Render/RenderDetailActions.php
	app/code/local/Tim/Recommendation/Helper/Data.php
	app/code/local/Tim/Recommendation/controllers/Adminhtml/CommentsReportController.php
	app/code/local/Tim/Recommendation/controllers/Adminhtml/OpinionReportController.php
	app/design/frontend/default/default/template/tim/recommendation/add/left.phtml
	app/design/frontend/default/default/template/tim/recommendation/profile/content.phtml
	app/design/frontend/default/default/template/tim/recommendation/profile/left.phtml
	app/design/frontend/default/default/template/tim/recommendation/view/left.phtml

+++ app/code/local/Tim/Recommendation/Helper/Data.php
@@ -38,7 +38,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerUserTypeId($customerId)
     {
-        $user = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $user = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $userTypeId = $user->getUserType();
 
         return $userTypeId;
@@ -63,7 +63,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getRecommendationMd5($recomId)
     {
-        $object = Mage::getModel('tim_recommendation/recommendation')->load($recomId);
+        $object = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId);
         $md5hash = $object->getMd5();
 
         return $md5hash;
@@ -76,7 +76,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getSiteUrl()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $siteUrl = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getWww();
         return $siteUrl;
     }
@@ -94,9 +94,9 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     {
         if (empty($customerId)) {
             $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
-            $nick = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getNick();
+            $nick = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id')->getNick();
         } else {
-            $nick = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getNick();
+            $nick = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id')->getNick();
         }
         return $nick;
     }
@@ -108,7 +108,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerBanner()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $banner = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getAd();
         if (!empty($banner)) {
             $banner = Mage::getBaseUrl(Mage_Core_Model_Store::URL_TYPE_WEB) . $banner;
@@ -125,7 +125,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerAvatar()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $avatar = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getAvatar();
         if (!empty($avatar)) {
             $avatar = Mage::getBaseUrl(Mage_Core_Model_Store::URL_TYPE_WEB) . $avatar;
@@ -142,7 +142,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerDescription()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $siteUrl = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getDescription();
         return $siteUrl;
     }
@@ -155,7 +155,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerName($customerId)
     {
-        $customer = Mage::getModel('customer/customer')->load($customerId);
+        $customer = Mage::getModel('customer/customer')->load((int) $customerId);
         $name = $customer->getFirstname() . ' ' . $customer->getLastname();
 
         return $name;
@@ -169,7 +169,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerNickname($customerId)
     {
-        $recommendationUser = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $recommendationUser = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $nickname = $recommendationUser->getNick();
 
         return $nickname;
@@ -183,9 +183,9 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerNameOrNick($customerId)
     {
-        $customerName = $this->getCustomerNickname($customerId);
+        $customerName = $this->getCustomerNickname((int) $customerId);
         if (empty($customerName)) {
-            $customerName = $this->getCustomerName($customerId);
+            $customerName = $this->getCustomerName((int) $customerId);
         }
 
         return $customerName;
@@ -199,7 +199,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getUserTypeName($userTypeId)
     {
-        $userType = Mage::getModel('tim_recommendation/userType')->load($userTypeId, 'user_type_id');
+        $userType = Mage::getModel('tim_recommendation/userType')->load((int) $userTypeId, 'user_type_id');
         $userTypeName = $userType->getName();
 
         return $userTypeName;
@@ -214,7 +214,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getOpinionMediaPath($opinionId)
     {
         $collection = Mage::getModel('tim_recommendation/media')->getCollection();
-        $collection->addFieldToFilter('recom_id', $opinionId);
+        $collection->addFieldToFilter('recom_id', (int) $opinionId);
         $data = $collection->getData();
         $mediaPaths = array();
         foreach ($data as $item) {
@@ -240,7 +240,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
         $opinionCollection->addFieldToSelect('user_id');
         $opinionCollection->addFieldToFilter('acceptance', 1);
         if ($customerId) {
-            $opinionCollection->addFieldToFilter('user_id', $customerId);
+            $opinionCollection->addFieldToFilter('user_id', (int) $customerId);
         }
         $qty = count($opinionCollection->getData());
 
@@ -345,7 +345,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function checkForOpinionComment($recomId)
     {
-        $recommendationRow = Mage::getModel('tim_recommendation/recommendation')->load($recomId);
+        $recommendationRow = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId);
         if ($parent = $recommendationRow->getParent()) {
             return $parent;
         } else {
@@ -440,7 +440,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionPoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/add_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -451,7 +451,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionImagePoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/image_to_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -462,7 +462,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionMoviePoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/movie_to_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -473,7 +473,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddComentPoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/add_comment');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -531,8 +531,8 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getUserScore($customerId)
     {
-        $customer = Mage::getModel('customer/customer')->load($customerId);
-        $user = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $customer = Mage::getModel('customer/customer')->load((int) $customerId);
+        $user = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $customerPoints = $user->getPoints();
         $customerLevel = $user->getLevel();
         $userLevelsClient = $this->getUserLevelClient();
@@ -579,7 +579,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getImages($recomId)
     {
-        $opinionMedia = $this->getOpinionMediaPath($recomId);
+        $opinionMedia = $this->getOpinionMediaPath((int) $recomId);
         $images = array();
         foreach ($opinionMedia as $key => $value) {
             if ($key === 'url/youtube') {
@@ -599,8 +599,8 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     {
         //check on acceptance for opinion or comment
         if (!$recommendationModel->getAcceptance()) {
-            $recomId = $recommendationModel->getRecomId();
-            $userId = $recommendationModel->getUserId();
+            $recomId = (int) $recommendationModel->getRecomId();
+            $userId = (int) $recommendationModel->getUserId();
             $userModel = Mage::getModel('tim_recommendation/user')->load($userId, 'customer_id');
             $mediaModel = Mage::getModel('tim_recommendation/media')->load($recomId, 'recom_id');
             $opinionOrComment = $this->checkOpinionOrComment($recomId);
@@ -636,7 +636,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     function checkOpinionOrComment($recomId)
     {
-        $parent = Mage::getModel('tim_recommendation/recommendation')->load($recomId)->getParent();
+        $parent = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId)->getParent();
         if ($parent) {
             return 'comment';
         } else {
